requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
kubernetes==28.1.0
docker==6.1.3
GitPython==3.1.40
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# Testing and Quality
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
playwright==1.40.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Additional IDP Platform Features
django-celery-beat==2.5.0
django-extensions==3.2.3
django-filter==23.5
django-health-check==3.17.0
django-prometheus==2.3.1
drf-spectacular==0.27.0
flower==2.0.1

# Monitoring and Observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Security and Compliance
django-ratelimit==4.1.0
django-csp==3.7
cryptography==41.0.7

# Infrastructure and Cloud Tools
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0
pyyaml==6.0.1
jinja2==3.1.2
jsonschema==4.20.0

# API and Integration
requests==2.31.0
httpx==0.25.2
websockets==12.0

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.18.1 
